Catch dialog dismissal error that occurs in tests
In testing, the sync activity can end before the progress dialog is
dismissed in onTaskCompleted. In this case, dismissing the dialog will
cause android to complain that the view isn't attached to a window
manager (even though the window manager is still somehow accessible
through getWindew().getWindowManager()...).

Tried several solutions, settled on catching the error in code and
logging a warning.